third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGImageElement-svgdom-width-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'width' property of the SVGImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageElement.width.baseVal.value is 100
PASS imageElement.width.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
